requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
pyzmq==26.0.3
Werkzeug==3.0.3
WTForms==3.1.2
